move mocks to the framework test fixtures
