+++ modules/framework/galasa-parent/dev.galasa.framework.api.bootstrap/build.gradle
@@ -10,7 +10,7 @@ dependencies {
     implementation project(':dev.galasa.framework.api.common')
 
     testImplementation 'org.apache.commons:commons-lang3'
-    testImplementation project(':dev.galasa.framework').sourceSets.test.output
+    testImplementation(testFixtures(project(':dev.galasa.framework')))
     testImplementation(testFixtures(project(':dev.galasa.framework.api.common')))
 }
