+++ modules/framework/galasa-parent/dev.galasa.framework.api.secrets/build.gradle
@@ -10,6 +10,7 @@ dependencies {
     implementation project(':dev.galasa.framework.api.beans')
     implementation project(':dev.galasa.framework.api.common')
     testImplementation(testFixtures(project(':dev.galasa.framework.api.common')))
+    testImplementation(testFixtures(project(':dev.galasa.framework')))
 }
 
 // Note: These values are consumed by the parent build process
