+++ modules/framework/galasa-parent/dev.galasa.framework.api.users/build.gradle
@@ -19,6 +19,7 @@ dependencies {
     
     testImplementation(testFixtures(project(':dev.galasa.framework.api.common')))
     testImplementation(testFixtures(project(':dev.galasa.framework.auth.spi')))
+    testImplementation(testFixtures(project(':dev.galasa.framework')))
 }
 
 // Note: These values are consumed by the parent build process
