+++ modules/framework/galasa-parent/dev.galasa.framework.api/build.gradle
@@ -10,7 +10,7 @@ dependencies {
     implementation 'org.apache.felix:org.apache.felix.bundlerepository'
     implementation 'org.osgi:org.osgi.service.cm'
 
-    testImplementation project(':dev.galasa.framework').sourceSets.test.output
+    testImplementation(testFixtures(project(':dev.galasa.framework')))
 }
 
 // Note: These values are consumed by the parent build process
