+++ modules/framework/galasa-parent/dev.galasa.framework.auth.spi/build.gradle
@@ -25,9 +25,9 @@ dependencies {
     testFixturesImplementation 'dev.galasa:dev.galasa.wrapping.io.grpc.java'
     testFixturesImplementation 'org.assertj:assertj-core:3.16.1'
     testFixturesImplementation(project(':dev.galasa.framework'))
+    testFixturesImplementation(testFixtures(project(':dev.galasa.framework')))
     testFixturesImplementation(project(':dev.galasa.framework.api.beans'))
     testFixturesImplementation(testFixtures(project(':dev.galasa.framework.api.common')))
-    testImplementation project(':dev.galasa.framework').sourceSets.test.output
     testFixturesImplementation 'javax.validation:validation-api'
 }
