+++ modules/framework/galasa-parent/dev.galasa.framework.k8s.controller/build.gradle
@@ -11,7 +11,7 @@ dependencies {
     implementation 'dev.galasa:dev.galasa.wrapping.io.kubernetes.client-java'
     implementation 'io.prometheus:simpleclient'
 
-    testImplementation project(':dev.galasa.framework').sourceSets.test.output
+    testImplementation(testFixtures(project(':dev.galasa.framework')))
 }
 
 // Note: These values are consumed by the parent build process
